ft_printf modification to include fd

+++ src/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: julauren <julauren@student.42angouleme.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/19 11:39:12 by julauren          #+#    #+#             */
-/*   Updated: 2025/12/12 15:32:46 by julauren         ###   ########.fr       */
+/*   Updated: 2026/06/02 11:07:56 by julauren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-static size_t	ft_intlen(int n)
+static size_t	ft_intlen(long n)
 {
 	int	i;
 
@@ -27,7 +27,7 @@ static size_t	ft_intlen(int n)
 	return (i);
 }
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long n)
 {
 	size_t	len;
 	char	*str;
